refactor(dashboard): extract telemetry domain (A1 phase 2 cont.)
- dashboard-telemetry.ts: TelemetrySource/Entry/Response types,
  DashboardCache* types, decodeTelemetry*/decodeDashboardCache* helpers,
  fetchTelemetry/fetchTelemetrySummary/fetchDashboardCacheStats
- dashboard.ts: facade re-export only; consumers unchanged
- typecheck green (tsc --noEmit)
